Optimized and Cleaned Code, Changed Object Colors
Optimized the code wherever possible, made code clean, changed powerup and food prefabs color so that they are more eye pleasing.
