+++ Co-Op-Snake-2D/Assets/Prefabs/Foods/Mass Burner.prefab
@@ -75,7 +75,7 @@ SpriteRenderer:
   m_SortingLayer: 0
   m_SortingOrder: -1
   m_Sprite: {fileID: 21300000, guid: 162c961f79f03a7419bae8455ff18508, type: 3}
-  m_Color: {r: 1, g: 0.8, b: 0.6, a: 1}
+  m_Color: {r: 0.745283, g: 0.26886433, b: 0.23483442, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
