+++ Co-Op-Snake-2D/Assets/Prefabs/Foods/Mass Gainer.prefab
@@ -75,7 +75,7 @@ SpriteRenderer:
   m_SortingLayer: 0
   m_SortingOrder: -1
   m_Sprite: {fileID: 21300000, guid: 162c961f79f03a7419bae8455ff18508, type: 3}
-  m_Color: {r: 1, g: 0.92156863, b: 0.5176471, a: 1}
+  m_Color: {r: 0.9870754, g: 1, b: 0.107547104, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
