+++ Co-Op-Snake-2D/Assets/Prefabs/PowerUps/Score Boost Power Up.prefab
@@ -75,7 +75,7 @@ SpriteRenderer:
   m_SortingLayer: 0
   m_SortingOrder: -1
   m_Sprite: {fileID: 21300000, guid: 37616095ebf53744c950962f33663458, type: 3}
-  m_Color: {r: 0.68235296, g: 0.8745098, b: 0.96862745, a: 1}
+  m_Color: {r: 0.40405837, g: 0.895941, b: 0.9433962, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
