+++ Co-Op-Snake-2D/Assets/Prefabs/PowerUps/Shield Power Up.prefab
@@ -75,7 +75,7 @@ SpriteRenderer:
   m_SortingLayer: 0
   m_SortingOrder: -1
   m_Sprite: {fileID: 21300000, guid: 37616095ebf53744c950962f33663458, type: 3}
-  m_Color: {r: 0.69411767, g: 0.6117647, b: 0.8509804, a: 1}
+  m_Color: {r: 0.90471286, g: 0.22584549, b: 0.9811321, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
