+++ Co-Op-Snake-2D/Assets/Prefabs/PowerUps/Speed Up Power Up.prefab
@@ -75,7 +75,7 @@ SpriteRenderer:
   m_SortingLayer: 0
   m_SortingOrder: -1
   m_Sprite: {fileID: 21300000, guid: 37616095ebf53744c950962f33663458, type: 3}
-  m_Color: {r: 1, g: 0.4117647, b: 0.38039216, a: 1}
+  m_Color: {r: 0.8301887, g: 0.38703802, b: 0.36340332, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
