+++ Co-Op-Snake-2D/Assets/Scenes/CoOpMode.unity
@@ -405,7 +405,7 @@ Camera:
   m_Enabled: 1
   serializedVersion: 2
   m_ClearFlags: 2
-  m_BackGroundColor: {r: 0.31399074, g: 0.6603774, b: 0.64039356, a: 0}
+  m_BackGroundColor: {r: 0.50980395, g: 0.6431373, b: 0.63529414, a: 0}
   m_projectionMatrixMode: 1
   m_GateFitMode: 2
   m_FOVAxisMode: 0
