+++ Co-Op-Snake-2D/Assets/Scripts/Game Control/GamePauseController.cs
@@ -1,4 +1,3 @@
-using System;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
@@ -12,17 +11,17 @@ public class GamePauseController : MonoBehaviour
 
     private void Awake()
     {
-        // Add listeners to buttons for corresponding methods.
-        resumeButton.onClick.AddListener(ResumeLevel); // Listener for the Resume button.
-        mainMenuButton.onClick.AddListener(GoToMainMenu); // Listener for the Main Menu button.
+        resumeButton.onClick.AddListener(ResumeLevel);
+        mainMenuButton.onClick.AddListener(GoToMainMenu);
     }
+
     private void ResumeLevel()
     {
         SoundManager.Instance.PlayEffect(SoundType.LevelPause);
-        // Activate the GamePause prefab
         gamePausePrefab.SetActive(false);
         Time.timeScale = 1f;
     }
+
     private void GoToMainMenu()
     {
         Time.timeScale = 1f;
@@ -33,7 +32,6 @@ public class GamePauseController : MonoBehaviour
     {
         SoundManager.Instance.PlayEffect(SoundType.LevelPause);
         Time.timeScale = 0f;
-        // Activate the GamePause prefab
         gamePausePrefab.SetActive(true);
     }
 }
