+++ Co-Op-Snake-2D/Assets/Scripts/Gameplay Mechanics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a5aebf76e18c6234aaa19da8e2db3050
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
