+++ Co-Op-Snake-2D/Assets/Scripts/Lobby and UI Management.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 19aa1557733397b4485b30356d3ac878
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
