+++ Co-Op-Snake-2D/Assets/Scripts/Managers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 88adc7b083dfd984496734b154f891de
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
