+++ Co-Op-Snake-2D/Assets/Scripts/Scene Management.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a9bc9b58db8f2c849b02d447f73edce1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
