+++ Co-Op-Snake-2D/Assets/Scripts/Scene Management/LevelLoader.cs
@@ -1,0 +1,31 @@
+using UnityEngine;
+using UnityEngine.UI;
+
+[RequireComponent(typeof(Button))]
+public class LevelLoader : MonoBehaviour
+{
+    private Button button;
+    public string buttonType;
+
+    private void Awake()
+    {
+        button = GetComponent<Button>();
+        button.onClick.AddListener(OnClick);
+    }
+
+    private void OnClick()
+    {
+        switch (buttonType)
+        {
+            case "Level":
+                SceneManagerUtility.LoadScene(transform.GetSiblingIndex() + 1);
+                break;
+            case "Back":
+                SceneManagerUtility.LoadMainMenu();
+                break;
+            default:
+                Debug.LogWarning("Unknown button type: " + buttonType);
+                break;
+        }
+    }
+}
